+++ examples/project/reflex-project-skeleton/default.nix
@@ -1,0 +1,1 @@
+import ((import <nixpkgs> {}).fetchFromGitHub (builtins.fromJSON (builtins.readFile ./github.json)))